feat: Implement Build, Targeted-Repair, and Retry Pipeline
This commit implements a new, highly efficient "Build-First" workflow for the AI website generator, as specified in the latest user requirements. This replaces the previous proactive analysis with a reactive, build-and-repair loop to significantly improve speed and cost-effectiveness.

The new workflow is as follows:
1.  **Bulk Generation:** All code is generated in a single pass without pre-build analysis.
2.  **Build, Repair, Retry Loop:** The system attempts to build the project up to three times. If a build fails, a targeted, syntax-focused AI repair is attempted on the specific file and line that caused the error.
3.  **Deployment:** The system proceeds to deployment only after a successful build.

Key changes:
- `agent/deployer.py` has been refactored to include the 3-strike build-and-retry loop.
- `agent/error_handler.py` is now responsible for parsing build errors and orchestrating targeted, syntax-focused fixes.
- `agent/enhanced_llm_service.py` and `agent/llm_service.py` have been simplified to remove all pre-build critic analysis, ensuring a fast initial code generation phase.
